refacto: move consumer logic to abstract consumer
