+++ bizon/engine/queue/adapters/kafka/queue.py
@@ -5,7 +5,8 @@ from kafka import KafkaProducer
 from loguru import logger
 
 from bizon.destination.destination import AbstractDestination
-from bizon.engine.queue.queue import QUEUE_TERMINATION, AbstractQueue, QueueMessage
+from bizon.engine.queue.queue import AbstractQueue
+from bizon.engine.queue.config import QUEUE_TERMINATION, QueueMessage
 
 from .config import KafkaConfigDetails
 from .consumer import KafkaConsumer_
