+++ bizon/engine/queue/adapters/python_queue/queue.py
@@ -6,11 +6,10 @@ from typing import Union
 from loguru import logger
 
 from bizon.destination.destination import AbstractDestination
+from bizon.engine.queue.config import QUEUE_TERMINATION, QueueMessage
 from bizon.engine.queue.queue import (
-    QUEUE_TERMINATION,
     AbstractQueue,
     AbstractQueueConsumer,
-    QueueMessage,
 )
 from bizon.source.models import SourceIteration
 from bizon.transform.transform import Transform
