+++ bizon/engine/queue/adapters/rabbitmq/queue.py
@@ -5,7 +5,8 @@ from loguru import logger
 
 from bizon.destination.destination import AbstractDestination
 from bizon.engine.pipeline.consumer import AbstractQueueConsumer
-from bizon.engine.queue.queue import QUEUE_TERMINATION, AbstractQueue, QueueMessage
+from bizon.engine.queue.queue import AbstractQueue
+from bizon.engine.queue.config import QUEUE_TERMINATION, QueueMessage
 
 from .config import RabbitMQConfigDetails
 from .consumer import RabbitMQConsumer
